requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
Faker==37.1.0
geographiclib==2.0
python-decouple==3.8
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pip==25.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2